fix: 不允许删除系统管理员用户以及角色

+++ src/main/java/com/linck/management/system/controller/SysRoleController.java
@@ -3,6 +3,7 @@ package com.linck.management.system.controller;
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.linck.management.common.api.Result;
+import com.linck.management.common.exception.BizException;
 import com.linck.management.common.model.dto.IdDto;
 import com.linck.management.common.model.dto.IdsDto;
 import com.linck.management.common.model.vo.ListWithPage;
@@ -71,6 +72,9 @@ public class SysRoleController {
      */
     @DeleteMapping
     public Result<String> remove(@RequestBody @Validated IdsDto idsDto) {
+        if (idsDto.getIds().contains(1L)) {
+            throw new BizException("系统管理员不允许删除");
+        }
         sysRoleService.removeBatchByIds(idsDto.getIds());
         return Result.success();
     }
